% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tbl.R
\name{delete_entry}
\alias{delete_entry}
\title{Delete a row from a table}
\usage{
delete_entry(con = NULL, tbl = NULL, ...)
}
\arguments{
\item{con}{DBI connection object.}

\item{tbl}{Name of the target table.}

\item{...}{Named primary key column-value pairs for the row to delete.}
}
\value{
None.
}
\description{
Deletes a row from a table based on primary keys.
}
\examples{
\dontrun{
con <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
delete_entry(con, "my_table", id = 1)
}

}
