% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tbl.R, R/fct_search.R
\name{get_db}
\alias{get_db}
\title{Retrieve joined data from the database}
\usage{
get_db(con = NULL)

get_db(con = NULL)
}
\arguments{
\item{con}{A DBI connection object. The database connection used to query the data.}
}
\value{
A lazy data frame (tibble) representing the joined data. The query is not executed until explicitly requested by the user (e.g., with \code{dplyr::collect()}).
}
\description{
This function retrieves data by performing a series of \verb{INNER JOIN} operations across multiple tables in the database,
including \code{lab_measurement}, \code{analyte}, \code{lab_sample}, \code{lab_field_sample}, \code{field_sample}, \code{sites}, and \code{species}.
The result is a combined dataset containing related information from all these tables.
}
