% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tbl.R
\name{get_tbl}
\alias{get_tbl}
\title{Retrieve data from a specified table in the database}
\usage{
get_tbl(con = NULL, tbl = NULL)
}
\arguments{
\item{con}{A DBI connection object returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}. This connection is used to interact with the database.}

\item{tbl}{A character string specifying the name of the target table in the database.}
}
\value{
A data frame containing all rows and columns from the specified table.
If the table is empty or does not exist, an error or empty data frame will be returned.
}
\description{
This function reads and returns all the rows from a specified table in the database.
It uses \code{\link[DBI:dbReadTable]{DBI::dbReadTable()}} to extract the contents of the table and return them as a data frame.
}
